Moved everything into a top-level package named abm_cali
Updated the setup.py to set the distribution name to abm-cali; you
can bikeshed about the name, but this one was not already used on
PyPI (just 'cali' was).

Updated any internal imports to use relative imports within the
abm_cali package.
