+++ setup.py
@@ -1,9 +1,9 @@
 from setuptools import find_packages, setup
 
 setup(
-    name='src',
-    packages=find_packages(),
+    name='abm-cali',
     version='0.1.0',
+    packages=find_packages(),
     description='A set off calibration tools to be used to calibrate Brook and Hommes model params.',
     author='Simudyne',
     license='MIT',
